Fork branching infrastructure for Nebula examine mode
KV cache sequence forking, per-branch trail rendering with phase-offset
orbits, and WebSocket protocol for streaming fork tokens. UX is rough —
spans don't show branches, scrub is not branch-aware, fork trigger is a
hidden 'f' key. Plumbing first, interaction design next.
